+++ src/asp/Core/StereoSettings.h
@@ -119,7 +119,6 @@ namespace asp {
     double nodata_value;                    ///< Pixels with values less than or equal to this number are treated as no-data.
                                             //  This overrides the nodata values from input images.
     double nodata_pixel_percentage;         ///< Percentage of low-value pixels treated as no-data
-    double nodata_optimal_threshold_factor; ///< Pixels with values less than this factor times the optimal Otsu threshold are treated as no-data
     double nodata_stddev_thresh;            ///
     int    nodata_stddev_kernel;            ///< Kernel size of the nadata stddev calculation
     bool   skip_rough_homography;           ///< Use this if datum-based rough homography fails.